// openapi/collector/model/ctListQueryRequest.ts
/**
 * 
 * MPDC Collector API
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CtListQueryRequest { 
    apiKey: string;
    fetchMultimedia: Array<CtListQueryRequest.FetchMultimediaEnum>;
    indexInCineast: boolean;
    label: string;
    collectionName: string;
    from: string;
    to: string;
    listIds?: Array<number>;
    queryDelayMinutes?: number;
    includeHistory?: boolean;
}
export namespace CtListQueryRequest {
    export type FetchMultimediaEnum = 'UNKNOWN' | 'ANIMATED_IMAGE' | 'IMAGE' | 'VIDEO';
    export const FetchMultimediaEnum = {
        Unknown: 'UNKNOWN' as FetchMultimediaEnum,
        AnimatedImage: 'ANIMATED_IMAGE' as FetchMultimediaEnum,
        Image: 'IMAGE' as FetchMultimediaEnum,
        Video: 'VIDEO' as FetchMultimediaEnum
    };
}